ref(cli): move overwriting trusted setup file to NodeCommand
